debugging

+++ __init__.py
@@ -275,6 +275,7 @@ def getUserID(email):
 
 def getUserInfo(user_id):
     user = session.query(User).filter_by(id=user_id).one()
+    print(user)
     return user
 
 
@@ -283,6 +284,7 @@ def createUser(login_session):
         name=login_session['username'],
         email=login_session['email'],
         picture=login_session['picture'])
+    print(newUser)
     session.add(newUser)
     session.commit()
     user = session.query(User).filter_by(email=login_session['email']).one()
@@ -356,16 +358,17 @@ def menuItemJson(category_id, menu_id):
 @app.route('/')
 def showCategorys():
     categorys = session.query(Category).order_by(asc(Category.name))
-    # if 'username' not in login_session:
-    #     return render_template('categoryList.html', categorys=categorys)
-    # else:
-    return render_template('publicCategory.html', categorys=categorys)
+    if 'username' not in login_session:
+        print('not in login')
+        return render_template('categoryList.html', categorys=categorys)
+    else:
+        return render_template('publicCategory.html', categorys=categorys)
 
 # Create a new Category
 @app.route('/category/new/', methods=['GET', 'POST'])
 def newCategory():
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         newCategory = Category(
             name=request.form['name'],
@@ -400,8 +403,8 @@ def editCategory(category_id):
 def deleteCategory(category_id):
     categoryToDelete = session.query(
         Category).filter_by(id=category_id).one()
-    # if 'username' not in login_session:
-    #     return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         session.delete(categoryToDelete)
         session.commit()
